+++ paimon-hive/paimon-hive-catalog/pom.xml
@@ -112,35 +112,6 @@ under the License.
             </exclusions>
         </dependency>
 
-        <!-- for hudi clone -->
-        <dependency>
-            <groupId>org.apache.hudi</groupId>
-            <artifactId>hudi-common</artifactId>
-            <version>${hudi.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.hudi</groupId>
-            <artifactId>hudi-flink</artifactId>
-            <version>${hudi.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.flink</groupId>
-            <artifactId>flink-core</artifactId>
-            <version>${paimon-flink-common.flink.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <dependency>
-            <groupId>org.apache.flink</groupId>
-            <artifactId>flink-table-common</artifactId>
-            <version>${paimon-flink-common.flink.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
         <!-- Test -->
 
         <dependency>
