+++ pom.xml
@@ -70,6 +70,7 @@ under the License.
         <module>paimon-arrow</module>
         <module>tools/ci/paimon-ci-tools</module>
         <module>paimon-open-api</module>
+        <module>paimon-hudi</module>
     </modules>
 
     <properties>
